Try a little harder to coerce SRIDs on writes.
If the column's or the geometry's SRID are -1 (unknown), then we may as
well assume that the incoming geometry's SRID should just be set to
whatever the column's is. If both of these SRIDs are known, then we
would have to a transformation, but since it's not all that easy with
ActiveRecord as it currently is, we'll just have to raise an exception
for the time being.